# services/api-gateway/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateAccountResponse:
    properties:
      account_id:
        type: string
    type: object
...
  models.CurrencyAccount:
    properties:
      account_name:
        type: string
      balance:
        type: number
      blocked_amount:
        type: number
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.GetBalanceResponse:
    properties:
      balance:
        type: number
    type: object
...
  models.Session:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      refresh_token_hash:
        type: string
      user_agent:
        type: string
      user_id:
        type: string
      user_ip:
        type: string
    type: object
...
  models.UserSession:
    properties:
      AccessToken:
        type: string
      RefreshToken:
        type: string
    type: object
host: localhost
info:
  contact: {}
  description: API для платежной системы
  title: Payment API
  version: "1.0"
paths:
  /account/{account_id}:
    get:
      description: Возвращает информацию о счете по его ID
      parameters:
      - description: Номер счета
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CurrencyAccount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить информацию о счете
      tags:
      - account
  /account/{account_id}/balance:
    get:
      description: Возвращает баланс счета по номеру счета
      parameters:
      - description: Номер счета
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetBalanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить баланс
      tags:
      - account
  /account/create:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Создает новый счет с указанной валютой и именем
      parameters:
      - description: Валюта
        enum:
        - USD
        - EUR
        - RUB
        in: formData
        name: currency
        required: true
        type: string
      - description: Имя счета
        in: formData
        name: account_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreateAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Создать счет
      tags:
      - account
  /account/get:
    get:
      description: Возвращает список всех счетов текущего пользователя
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CurrencyAccount'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить список счетов пользователя
      tags:
      - account
  /auth/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Валидирует входные данные и создаёт сессию(сохраняет токены в cookies)
      parameters:
      - description: User email
        in: formData
        name: email
        required: true
        type: string
      - description: User password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserSession'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Войти используя данные пользователя
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Удаляет текущую сессию(добавляет токен в Blacklist до истечения
        его срока жизни, удаляет данные о сессии из базы данных)
      parameters:
      - description: ID пользователя
        in: query
        name: userID
        type: string
      - description: ID сессии
        in: query
        name: sessionID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Выйти из текущей сессии
      tags:
      - auth
  /auth/logout-all:
    post:
      consumes:
      - application/json
      description: Удаляет все сессии текущего пользователя
      parameters:
      - description: ID пользователя
        in: query
        name: userID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Выйти со всех устройств
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Обновляет сессию(токены), для получения нового, не просроченного
        Access токена
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserSession'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Обновить сессию
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Регистрирует нового пользователя и создаёт для него сессию(сохраняет
        токены в cookies)
      parameters:
      - description: User name
        in: formData
        name: name
        required: true
        type: string
      - description: User email
        in: formData
        name: email
        required: true
        type: string
      - description: User password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserSession'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Зарегистрировать нового пользователя
      tags:
      - auth
  /auth/sessions:
    get:
      consumes:
      - application/json
      description: Возвращает все сессии пользователя
      parameters:
      - description: ID пользователя
        in: query
        name: userID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Session'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Получить все сессии пользователя
      tags:
      - auth
...
schemes:
- https
securityDefinitions:
  CookieAuth:
    description: Аутентификация через куки (токен автоматически отправляется сервером)
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"